Added adjustments
