+++ src/sass/_pagination.scss
@@ -56,3 +56,9 @@
 body {
   background: #0E0E0E;
 }
+
+.while-searching_text {
+  color: white;
+  font-family: Montserrat;
+  font-size: 36px;
+}
